WTA #31: Moved network related classes to separate package.
